fix: support proper note title during creation
The note title was not being properly handled during creation, appearing as
"Untitled" on HackMD.

HackMD API doc says:
As for now, the title of a note is derived from the content, thus the behavior
of "title" field may need further explanation.
If there's a H1 header in the "content" field, then it will be used as the title
of the note to be created.
If there's no H1 header in the "content" field, yet a title is assigned in YAML
metadata section (---\ntitle: abc\n---), then it will be the title of the note.
If the "content" field is not included, then either look for the "title" field or
create the note as "Untitled" when it is absent.

This introduces temporary metadata duplication between "title" and
"hackmd.title". This change only handles title sync during note creation,
existing notes sync should be fixed separately.